small gui refactor done
