+++ python/gui/UI_demo/run_UIdemo.py
@@ -1,6 +1,7 @@
 import sys
 from PyQt5.QtWidgets import QApplication
-from windows.mainwindow import MainWindow
+
+from python.gui.windows.mainwindow import MainWindow
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
