+++ python/gui/widgets/tabs/logtab.py
